requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.2.3
bcrypt==4.0.1
